requirements.txt
telethon>=1.34.0
python-dotenv
pysocks